Create method.stratifiedNNLS.R
A single algorithm in the library ('SL.stratify' in the example) is used to data adaptively identify subgroups in the observed data. The other library algorithms are fit with the full training data, but weights can vary by subgroup. This differs from the subsemble method (https://github.com/ledell/subsemble) in that an overall predictor is fit and the weights change by subgroup, whereas subsemble estimates subset specific predictors.